Update Vagrantfile
This is my final test with the entire project. I didn't feel that there is a need for port forward configuration for the vagrant file. Because we are using the bridge adapter. When we are working with a bridge adapter we don't need to define port forward configuration in the vagrant file or manually with the virtual box.

Memory size 4096 is enough to complete all the necessary tasks related to the project. It is optional to increase the memory size 8196.